server/requirements.txt
Django==5.1.6
django-cors-headers
djangorestframework
